% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shiny_util.R
\name{ensure_reactable}
\alias{ensure_reactable}
\title{Flexible specification of parameters to shiny apps.}
\usage{
ensure_reactable(item, env_name = NULL)
}
\arguments{
\item{item}{Thing to ensure is "reactable".}

\item{env_name}{Name of variable in environment to retrieve value from, if \code{item} is NULL.}
}
\description{
Make a function to retrieve a parameter from an environment if not explicitly specified, and if it is not already such a function.
}
