% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/grid_util.R
\name{heatmap_grob}
\alias{heatmap_grob}
\title{A grob for heatmaps.}
\usage{
heatmap_grob(
  data,
  signed = TRUE,
  legend_title = "",
  vp_name = NULL,
  to = NULL,
  raster = FALSE
)
}
\arguments{
\item{data}{Matrix of data to display.}

\item{signed}{Should a signed palette be used?}

\item{legend_title}{Title for legend grob.}

\item{vp_name}{Viewport name for heatmap grob. This is useful if you will need to refer to the grob later, for example for interaction.}

\item{to}{Heatmap scale will include this value if given.}
}
\value{
A list containing a heatmap grob and associated legend grob.
}
\description{
A grob for heatmaps.
}
