% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/samesum.R
\name{samesum_log2_cpm}
\alias{samesum_log2_cpm}
\title{Normalized log2 CPM using "scalesum" method}
\usage{
samesum_log2_cpm(
  x,
  scale = 0.25,
  target_sum = NA,
  tol = 1e-09,
  verbose = FALSE
)
}
\description{
Applies samesum_log2_norm, then adds an adjustment to produce values that can be treated as log2 Counts Per Million (CPM).
}
\details{
The result will have the property that \code{mean(colSums(2**lcpm-2**attr(lcpm,"zero")))} equals one million.
}
