% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shiny_plots.R
\name{shiny_biplot}
\alias{shiny_biplot}
\title{Shiny wrapper for plot_biplot}
\usage{
shiny_biplot(x, sample_labels = NULL, feature_labels = NULL,
  n_features = 20, balance = 0.25, text_size = 0.025, prefix = "")
}
\arguments{
\item{x}{Transformed matrix.}

\item{sample_labels}{Sample labels.}

\item{feature_labels}{Feature labels.}

\item{n_features}{Number of features to label, by default.}

\item{balance}{Relative scaling of samples and features, by default.}

\item{text_size}{Text size in plot.}

\item{prefix}{Prefix for variables in environment for composable shiny app.}
}
\description{
Shiny wrapper for plot_biplot
}
