% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shiny_report.R
\name{shiny_filter}
\alias{shiny_filter}
\title{Shiny component to filter rows and columns a matrix.}
\usage{
shiny_filter(
  y,
  counts = NULL,
  sample_labels = NULL,
  feature_labels = NULL,
  prefix = ""
)
}
\arguments{
\item{y}{Transformed matrix.}

\item{counts}{Original count matrix if available.}

\item{sample_labels}{Sample labels.}

\item{feature_labels}{Feature labels.}

\item{prefix}{Prefix for variables in environment for composable shiny app.}
}
\description{
Shiny component to filter rows and columns a matrix.
}
