% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shiny_plots.R
\name{shiny_heatmap}
\alias{shiny_heatmap}
\title{Shiny wrapper for plot_heatmap}
\usage{
shiny_heatmap(y, sample_labels = NULL, feature_labels = NULL, prefix = "")
}
\arguments{
\item{y}{Transformed matrix.}

\item{sample_labels}{Sample labels.}

\item{feature_labels}{Feature labels.}

\item{prefix}{Prefix for variables in environment for composable shiny app.}
}
\description{
Shiny wrapper for plot_heatmap
}
