% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shiny_plots.R
\name{shiny_mds_plot}
\alias{shiny_mds_plot}
\title{Shiny wrapper for limma's MDS plot}
\usage{
shiny_mds_plot(x, sample_labels = NULL, prefix = "")
}
\arguments{
\item{x}{Transformed matrix.}

\item{sample_labels}{Sample labels.}

\item{prefix}{Prefix for variables in environment for composable shiny app.}
}
\description{
Shiny wrapper for limma's MDS plot
}
