% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shiny_report.R
\name{shiny_report}
\alias{shiny_report}
\title{Shiny report.}
\usage{
shiny_report(y, counts = NULL, prefix = "")
}
\arguments{
\item{y}{A matrix of exprssion levels, such as a transformed counts matrix.}

\item{counts}{Optional. Original counts.}

\item{prefix}{Optional, to fascilitate use as a component of a larger Shiny app. Inputs and outputs are given this prefix.}
}
\value{
A shiny.appobj.

Used interactively, the shiny report runs immediately. Otherwise it can be launched by print()-ing it. A call to this function can also be the last line in an app.R file in a Shiny app directory.
}
\description{
Produce an interactive Shiny report showing diagnostic plots of transformed counts.
}
\examples{

y <- varistran::vst(counts)
varistran::shiny_report(y, counts)

}
\author{
Paul Harrison
}

