% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shiny_plots.R
\name{shiny_stability}
\alias{shiny_stability}
\title{Shiny wrapper for plot_stability}
\usage{
shiny_stability(y, x = NULL, design = NULL, bins = 20, prefix = "")
}
\arguments{
\item{y}{Transformed matrix.}

\item{x}{Untransformed matrix.}

\item{design}{Design matrix.}

\item{bins}{Number of bins.}

\item{prefix}{Prefix for variables in environment for composable shiny app.}
}
\description{
Shiny wrapper for plot_stability
}
