% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shiny_report.R
\name{shiny_vst}
\alias{shiny_vst}
\title{Shiny component to describe the output of a call to vst.}
\usage{
shiny_vst(y = NULL, counts = NULL, sample_labels = NULL, prefix = "")
}
\arguments{
\item{y}{Transformed matrix.}

\item{counts}{Counts matrix to transform.}

\item{sample_labels}{Sample labels.}

\item{prefix}{Prefix for variables in environment for composable shiny app.}
}
\description{
Shiny component to describe the output of a call to vst.
}
