% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/grid_util.R
\name{shrinktext_grob}
\alias{shrinktext_grob}
\title{A collection of horizontal labels which shrink if they get too small.}
\usage{
shrinktext_grob(label, x, y, just, max_width = 20, ...)
}
\arguments{
\item{label}{Vector of labels.}

\item{x}{Vector of x positions.}

\item{y}{Vector of y positions.}

\item{just}{Text justification.}

\item{max_width}{Maximum width of labels (defaults to units of lines).}

\item{...}{Further arguments to \code{grob}.}
}
\description{
A collection of horizontal labels which shrink if they get too small.
}
