% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/grid_util.R
\docType{data}
\name{signed_colors}
\alias{signed_colors}
\title{A red-blue palette suitable for signed data.}
\format{An object of class \code{character} of length 256.}
\usage{
signed_colors
}
\description{
A red-blue palette suitable for signed data.
}
\keyword{datasets}
