% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/grid_util.R
\docType{data}
\name{unsigned_colors}
\alias{unsigned_colors}
\title{A palette from cool to hot.}
\format{An object of class \code{character} of length 256.}
\usage{
unsigned_colors
}
\description{
A palette from cool to hot.
}
\keyword{datasets}
