% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/grid_util.R
\name{varistran_grob}
\alias{varistran_grob}
\title{Make a grob better.}
\usage{
varistran_grob(grob, width = unit(1, "null"), height = unit(1, "null"),
  pad = 0)
}
\description{
Give a grob a margin, and size hints for layout, and make it print()-able.
}
