% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/grid_util.R
\name{vertical_shrinktext_grob}
\alias{vertical_shrinktext_grob}
\title{A collection of vertical labels which shrink if they get too small.}
\usage{
vertical_shrinktext_grob(label, x, y, just, ...)
}
\arguments{
\item{label}{Vector of labels.}

\item{x}{Vector of x positions.}

\item{y}{Vector of y positions.}

\item{just}{Text justification.}

\item{...}{Further arguments to \code{grob}.}
}
\description{
A collection of vertical labels which shrink if they get too small.
}
