% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/varistran.R
\name{vst_advice}
\alias{vst_advice}
\title{Advise how VST will transform data}
\usage{
vst_advice(what = "anscombe.nb", dispersion = NULL, cpm = FALSE,
  lib.size = NULL)
}
\arguments{
\item{what}{Either the output of a call to vst() or the name of a VST method (see vst() help).}

\item{dispersion}{As per vst().}

\item{cpm}{As per vst().}

\item{lib.size}{As per vst().}
}
\value{
A data frame giving an indication of how an average sample will be transformed.

The column "twofold_step" shows the step from the previous to current row. With a log2 transformation this would be uniformly 1, but with a VST and small counts the step is less than 1. This therefore provides advice on how compacted the VST is near zero-count, as compared to a log2 transformation.

Note that the results given are for an average sample. Where library sizes differ wildly, the VST may perform poorly.
}
\description{
Advise how VST will transform data
}
