% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shiny_util.R
\name{call_ui}
\alias{call_ui}
\title{Call object with "request" if it is callable.}
\usage{
call_ui(ui, request)
}
\description{
This is used to support older UI code which doesn't wrap UI in function(request) { ... }.
}

