% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/grid_util.R
\name{ordering_grob}
\alias{ordering_grob}
\title{orderingGrob}
\usage{
ordering_grob(ordering, transpose = FALSE, mirror = FALSE,
  hint_size = unit(5, "lines"))
}
\description{
Display the dendrogram of an ordering, as a grid grob.
}
\seealso{
\code{\link{make_ordering}}
}
