% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/heatmap.R
\name{plot_heatmap}
\alias{plot_heatmap}
\title{Plot a heatmap.}
\usage{
plot_heatmap(y, cluster_samples = FALSE, cluster_features = TRUE,
  sample_labels = NULL, feature_labels = NULL)
}
\arguments{
\item{y}{A matrix of expression levels, such as a transformed counts matrix.}

\item{cluster_samples}{Should samples (columns) be clustered?}

\item{cluster_features}{Should features (rows) be clustered?}

\item{sample_labels}{Names for each sample. If not given and y has column names, these will be used instead.}

\item{feature_labels}{Names for each feature. If not given and y has row names, these will be used instead.}
}
\value{
A grid grob. print()-ing this value will cause it to be displayed.

Additionally $info$row_order will contain row ordering and $info$col_order will contain column ordering.
}
\description{
Produces a heatmap as a grid grob.
}
\details{
Clustering is performed using the "seriation" package, and is approximately a Travelling Salesman Problem ordering. If there are many features (more than a couple of thousand) clustering may be slow.
}
\author{
Paul Harrison.
}

