% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/samesum.R
\name{samesum_log2_norm}
\alias{samesum_log2_norm}
\title{Normalized log2 counts using log2(count[i,j]*scale[j]+1)}
\usage{
samesum_log2_norm(x, scale = 0.25, tol = 1e-09, verbose = FALSE)
}
\description{
Normalized log2 counts using log2(count[i,j]*scale[j]+1)
}
