% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shiny_util.R
\name{shiny_plot}
\alias{shiny_plot}
\title{Shiny plot.}
\usage{
shiny_plot(callback, width = 500, height = 500, dlname = "plot",
  prefix = "")
}
\arguments{
\item{callback}{Function to produce the plot. Takes one argument, "env" (see composable_shiny_app).}
}
\value{
A composable shiny.appobj.
}
\description{
Convert a function that produces a plot into a Shiny app.
}
\author{
Paul Harrison
}

