% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shiny_report.R
\name{shiny_report}
\alias{shiny_report}
\title{Shiny report}
\usage{
shiny_report(y = NULL, counts = NULL, sample_labels = NULL,
  feature_labels = NULL, prefix = "")
}
\arguments{
\item{y}{A matrix of exprssion levels, such as a transformed counts matrix.}

\item{counts}{Original counts.}

\item{sample_labels}{Optional. Sample names.}

\item{feature_labels}{Optional. Feature names.}

\item{prefix}{Optional, to facilitate use as a component of a larger Shiny app. Inputs and outputs are given this prefix.}
}
\value{
A shiny.appobj.

Either y or counts or both must be given.

Used interactively, the shiny report runs immediately. Otherwise it can be launched by print()-ing it. A call to this function can also be the last line in an app.R file in a Shiny app directory.
}
\description{
Produce an interactive Shiny report showing diagnostic plots of transformed counts.
}
\examples{

# Generate some random data.
means <- runif(100,min=0,max=1000)
counts <- matrix(rnbinom(1000, size=1/0.01, mu=rep(means,10)), ncol=10)

y <- varistran::vst(counts)
if (interactive())
    varistran::shiny_report(y, counts)

}
\author{
Paul Harrison
}
